Consolidate legacy component detection utilities
