+++ packages/core/src/component/index.ts
@@ -1,10 +1,10 @@
 export * from "./component-collector";
 export * from "./component-collector-legacy";
 export * from "./component-detection";
+export * from "./component-detection-legacy";
 export * from "./component-flag";
 export * from "./component-id";
 export * from "./component-kind";
-export * from "./component-method";
 export * from "./component-name";
 export type * from "./component-semantic-node";
 export * from "./component-wrapper";
